EncodingsDialog: add Up and Down buttons to order chosen encodings
I used glade to edit the .ui file. Keep the changes so that editing
again the file with glade in the future will produce a smaller diff.

The Up and Down buttons work only when one encoding is selected. The
buttons sensitivity is correctly updated, even for the first/last
encodings (the Up button is insensitive for the first item, and the Down
button is insensitive for the last item).